Moved addUser method to login presenter
