+++ src/codeu/chat/client/android/app/src/main/java/com/google/codeu/chatme/controller/Controller.java
@@ -9,33 +9,17 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
+
+@Deprecated
+/**
+ * Functions defined in this class are being moved to presenter classes to follow MVP
+ * design pattern
+ */
 public class Controller {
 
     private static final String TAG = Controller.class.getName();
 
     private static DatabaseReference mRootRef = FirebaseDatabase.getInstance().getReference();
-
-    /**
-     * Saves a new user to Firebase real-time database
-     *
-     * @param id   id of {@link User}
-     * @param name display name of {@link User}
-     */
-    public static void addUser(final String id, String name) {
-        User newUser = new User(name);
-
-        mRootRef.child("users").child(id).setValue(newUser, new DatabaseReference.CompletionListener() {
-
-            @Override
-            public void onComplete(DatabaseError databaseError, DatabaseReference databaseReference) {
-                if (databaseError != null) {
-                    Log.w(TAG, "addUser:failure " + databaseError.getMessage());
-                } else {
-                    Log.i(TAG, "addUser:success " + id);
-                }
-            }
-        });
-    }
 
     /**
      * Saves a new conversation to Firebase real-time database
